Added first pass at the cross-platform framwork.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@353 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/ev/unix/ev_UnixKeyboard.cpp
@@ -312,7 +312,7 @@ static EV_EditBits s_mapVirtualKeyCodeToNVK(gint nVirtKey)
 	// these are referenced by NVK_ symbol so that the cross
 	// platform code can properly refer to them.
 	
-	UT_ASSERT(nVirtKey <= NrElements(s_Table_NVK));
+	UT_ASSERT(nVirtKey <= (gint)NrElements(s_Table_NVK));
 
 	return s_Table_NVK[nVirtKey];
 }
